Increased the limit for the "listen" section to 10000 and for the "global" section to 100000.
By default, the listening section is set to 2000. You may need a higher value to service a large number of client connections.
The "maxconn" parameter specifies the maximal number of concurrent connections that will be sent to this server. Proxies will stop accepting connections when this limit is reached.

Now, this value is set to 10000.

The "ulimit-n" parameter is automatically adjusted according to this value.