Increase the range of local ports available for outgoing connections.
When you are facing simultaneous 30k connections, you will encounter the problem of “port exhaustion”. It is resulted from the fact that each reverse proxied connection will  occupy an available port of a local IP.
The default IP range that is available for outgoing connections is set " 32768    60999". In other words, we only have 28k ports available for one IP.
This is may not enough. We increase this range by modify /etc/sysctl.conf file.

Now, this value is set to "1024 65535" by default. This is about 60k ports available for one IP address.

p.s.
The ultimate solution to this port exhaustion issue is to increase the number of available IPs.