Set umask before installing pip packages.
This is useful, for example, when installing on systems that have a very restrictive umask by default (e.g., "0077") and you want to pip install packages which are to be used by all users. Note that this requires you to specify desired umask mode as an octal string, (e.g., "0022").
https://docs.ansible.com/ansible/latest/modules/pip_module.html