Add RHEL/CentOS 8 compatibility.
* Added "os_specific_packages" variable with a list of requarements for specific RHEL Linux versions.

* Added chronyd (ntp) service.
In RHEL Linux 8, the ntp package is no longer supported and it is implemented by the chronyd (a daemon that runs in user-space) which is provided in the chrony package.

* Added the glibc-langpack installation instead of generating a locales (for RHEL 8 only).
Variable: "glibc_langpack" in the RedHat.yml inventory file.

* Added disable postgresql module for RHEL 8.
The version of PostgreSQL available on CentOS 8 / RHEL 8 officially maintained modular repository is 10 and 9.6. We’ll add PostgreSQL Yum Repository (by default) which hosts all recent releases.

* Install pgbouncer package with the "--disablerepo AppStream" parameter. Since the AppStream repository (for today) does not have the required version of python-psycopg2. For RHEL 8 only.
Fixed dpendency roblem: package pgbouncer-1.12.0-1.rhel8.x86_64 requires python-psycopg2, but none of the providers can be installed.

Other:
* Fix parameters of the "package" module for HAProxy on RHEL Linux 8.
* Fix ansible-role-firewall: added a Boolean filter for disable-other-firewalls task.

Tested on CentOS Linux release 8.0.1905