Improved postgresql instance status check for postgresql_exists='true'.
This playbook support the deployment of cluster over already existing and running PostgreSQL.
There is no need to stop the postgresql instance before starting this playbook, as the playbook will automatically restart the postgresql after preparation for working in high availability cluster mode.

Currently, the postgresql instance will automatically started if it was stopped.