Add synchronous_mode_strict variable.
When it is absolutely necessary to guarantee that each write is stored durably on at least two nodes, enable synchronous_mode_strict in addition to the synchronous_mode. This parameter prevents Patroni from switching off the synchronous replication on the primary when no synchronous standby candidates are available.
See replication modes documentation for details:
https://patroni.readthedocs.io/en/latest/replication_modes.html#replication-modes