Add patroni_create_replica_methods variable.
For "pgbackrest" or "wal_e", and “basebackup” by default.

create_replica_methods: an ordered list of the create methods for turning a Patroni node into a new replica. “basebackup” is the default method; other methods are assumed to refer to scripts, each of which is configured as its own config item.

See custom replica creation methods documentation for further explanation:
https://patroni.readthedocs.io/en/latest/replica_bootstrap.html#building-replicas