fix issue with unbalanced quotes.
ansible-lint:

The error appears to be in '/root/postgresql_cluster/tasks/wal-g.yml': line 57, column 24, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

- name: "wal-g | copy "wal_g_bootstrap.sh" to /etc/patroni/"
                       ^ here
We could be wrong, but this one looks like it might be an issue with
unbalanced quotes. If starting a value with a quote, make sure the
line ends with the same set of quotes.