remove example ip address from cluster_vip.
In variable "cluster_vip" must specify the ip address for your subnet. 
If you leave the ip address that was specified as an example, then you may receive an error when trying to check the availability of the cluster ip address during deployment:
FAILED! => {"changed": false, "elapsed": 61, "msg": "Timeout when waiting for 10.128.64.145:22"}

The cluster ip address (VIP) is optional, you can leave the cluster_vip variable empty, or specify the ip address for your subnet. This VIP will be used for client access to databases in the cluster.