Check that /usr/bin/pgbouncer is exists (RHEL)
#38
The pgbouncer package from the postgrespro repository installs the binary file under the path /usr/sbin/pgbouncer, but the package from the pgdg repository for the RedHat family is installed under the path /usr/bin/pgbouncer.

Check for the presence of the file /usr/bin/pgbouncer, if it is not exist, create a symlink on /usr/sbin/pgbouncer.
This is necessary for compatibility with the standard pgbouncer package from the pgdg repository.