Wait for the new cluster to initialize before generating the custom pg_hba.conf
Before that, we generated the user pg_hba.conf file immediately after checking whether PostgreSQL is running and whether it accepts connections on the Master server.
When using the custom bootstrap method, the patroni service can overwrite our pg_hba.conf file after the actual initialization of the new cluster is completed.

Now we wait until the master writes the leader key in the DCS and when the Patroni node is running as the leader.