Generate virtual_router_id based on the last octet of cluster_vip.
Use the last octet of cluster_vip instead of id in keepalived_virtual_router_id variable.
VIP will not work if you specify the same virtual_router_id for different postgres clusters within the same subnet.

virtual_router_id is a numerical identifier for the Virtual Router instance. It must be the same and unique for all LVS Router systems participating in this Virtual Router.