PITR (custom bootstrap) - disable archive_command
Before running patroni, make sure the archive_command is disabled after restoring from a backup.

This is necessary to avoid conflicts in the archived log storage when archiving WALs. When multiple clusters try to send WALs to the same storage.
For example, when you make multiple clones of a cluster from one backup.