PITR: "disable_archive_command" and "keep_patroni_dynamic_json" variables
* disable_archive_command:

By default, we disable archive_command after restoring from a backup.
This is necessary to avoid conflicts in the archived log storage when archiving WALs. When multiple clusters try to send WALs to the same storage. For example, when you make multiple clones of a cluster from one backup.

You can change this parameter using `patronictl edit-config` after restore.
Or set `disable_archive_command: false` to not disable archive_command after restore.

* keep_patroni_dynamic_json:

Patroni nodes are dumping the state of the DCS options to disk upon for every change of the configuration into the file patroni.dynamic.json located in the Postgres data directory. The master (patroni leader) is allowed to restore these options from the on-disk dump if these are completely absent from the DCS or if they are invalid.

Patroni will override the parameters specified in vars/main.yml if there is a patroni.dynamic.json file exist in the backup.
Set keep_patroni_dynamic_json: false to remove patroni.dynamic.json after restore (if exists).

Issue #82