Add remove_cluster.yml
An Ansible script for easy undeployment of a PostgreSQL cluster while experimenting with different configurations.

Could also be called 'undeploy_cluster.yml'.

To prevent unintentionally undeployment 'remove_postgres' and 'remove_etcd' should be set to 'true'.

TODO: If it is possible to reuse the variable 'pgbackrest_conf global repo1-path' that will come in handy. Otherwise a new variable 'pgbackrest_repo1-path' could be made and that then used for the config option also.