add_pgnode: add "new_node" variable
Previously, it was necessary to add a new node (which you want to add to existing cluster) in the [replica] group and at the same time remove other nodes from the replica group from inventory in order to add_pgnode.yml playbook was executed specifically for the new server.
Now you do not need to do this, you just need to specify the variable new_node=true for the server that you are adding to the cluster and the playbook will be executed only on this server.