Integrate code improvements and add linters (#307)
* 🔧 chore(linters): add linters for YAML, Ansible, and Python files
This commit adds linters for YAML, Ansible, and Python files. The `.yamllint` file is updated to ignore the `.venv` directory and to enforce a minimum of one space between comments and content. The `.ansible-lint` file is updated to exclude the `.venv` directory. The `Makefile` is updated to include the new `linters.mak` file, which contains the targets for the linters. The linters are executed in the `lint` target.

* 🔧 chore(molecule.yml): remove commented out code and fix indentation
This commit removes commented out code and fixes indentation in the molecule.yml file. The commented out code is not needed and the indentation is inconsistent, which can make the file harder to read.

* 🚀 chore(.gitpod.Dockerfile): add bash-completion and remove redundant packages
The bash-completion package is added to enable tab completion in the terminal. The following packages are removed as they are either redundant or not required: sudo, nano, iproute2, curl, htop, gnupg, and lsb-release.

* 🎨 style(linters.mak): add lint target to Makefile
The lint target is added to the Makefile to run all linters at once. This improves the development workflow by allowing developers to run all linters with a single command.

* 🆕 chore(.flake8): add flake8 configuration file
A new flake8 configuration file has been added to the .config directory. This file specifies the directories and files to exclude from flake8 linting.

* 🚚 chore(linters.mak): move flake8 configuration file to .config directory
The flake8 configuration file was moved from the root directory to the .config directory to improve organization and avoid cluttering the root directory.

* 🔧 chore(linters.mak): remove comments and echo statements
🔧 chore(postgrespro): move vars file to vars directory
The comments and echo statements were removed from the linters.mak file to reduce noise in the output. The vars file was moved to the vars directory to improve organization and consistency with other roles.

* 📝 docs(CONTRIBUTING.md): update contributing guidelines
This commit updates the contributing guidelines to include instructions on how to lint the codebase using `make lint`. It also adds a step to the existing instructions to test with linters.

* 🔀 chore(converge.yml): remove redundant ansible.builtin prefix from include_vars module
The ansible.builtin prefix is redundant and can be removed from the include_vars module.

* 🔥 chore(molecule.mak): remove molecule-lint target
The molecule-lint target has been removed as it is no longer needed.