Implement DRY principle with a shared config.yml file for Molecule scenarios (#348)
* 🔧 chore(molecule): remove duplicated code and enable balancers in molecule.yml
The molecule configuration files were refactored to remove duplicated code and enable balancers in the molecule.yml files. The dependency and provisioner sections were removed as they were already defined in the config.yml file.

* 🔥 chore(molecule.yml): remove trailing newline at end of file
This commit removes the trailing newline at the end of the molecule.yml file. The newline is not necessary and can cause issues with some tools that expect files to end with a newline.