🔧 chore(Makefile): add .SHELLFLAGS to enable strict mode (#349)
The .SHELLFLAGS option is added to the Makefile to enable strict mode. This option sets the shell to exit immediately if any command exits with a non-zero status, and it also sets the shell to exit immediately if any pipeline in the script fails. This ensures that the Makefile is more robust and less error-prone.