🐛 fix(Dockerfile): replace hardcoded Python version with version specified in .config/python_version.config (#350)
✨ feat(Dockerfile): add .config/python_version.config file to specify Python version
✨ feat(Makefile): replace hardcoded Python version with version specified in .config/python_version.config
The Dockerfile now uses the Python version specified in the .config/python_version.config file instead of a hardcoded version. This allows for easier version management and flexibility. The .config/python_version.config file is a new file that specifies the Python version to be used. The Makefile has also been updated to use the Python version specified in the .config/python_version.config file.