Resolves Multiple Issues and Enhances Docker Setup (#352)
* 🐛 fix(Dockerfile): fix apt-get install command to properly install python version
The apt-get install command was not properly installing the desired python version due to the quotes being removed from the package names. This commit adds quotes to the package names to ensure that the correct version of python is installed.

* 🐛 fix(Dockerfile): use correct command to extract python version from config file
The command to extract the python version from the config file was incorrect. The cut command was used instead of cat to extract the value of the python version. The cut command was updated to extract the value of the python version from the config file.

* 🔧 chore(Dockerfile): add npm and ungit to the Dockerfile
The Dockerfile has been updated to include the installation of npm and ungit. This is useful for development purposes as it allows developers to easily visualize and manage their git repositories.

* 🔧 chore(Dockerfile): add tree package to improve development experience
The tree package has been added to the Dockerfile to improve the development experience by allowing developers to easily visualize the directory structure of the project.

* 🚀 feat(Dockerfile): add variables for npm and ungit versions and use them to install the specified versions
The Dockerfile now includes two variables, NPM_VERSION and UNGIT_VERSION, which are used to install the specified versions of npm and ungit respectively. This allows for easier version management and ensures that the correct versions are installed.