Refactor Project Structure and Makefile Targets (#357)
* 🔧 chore(config): add yamllint and ansible-lint configuration files
✨ feat(config): add make targets for testing and bootstrapping development environment
🔧 chore(config): move dev requirements to a separate file
🔧 chore(config): update make targets to include development environment
The commit adds configuration files for yamllint and ansible-lint. It also adds make targets for testing and bootstrapping the development environment. The development environment is now bootstrapped separately from the production environment. The make targets have been updated to include the development environment. The dev requirements have been moved to a separate file.

* 🔥 chore(workflows): remove Ansible-lint and Yamllint workflows
✨ feat(workflows): add linters workflow to run all linters and improve code quality
The Ansible-lint and Yamllint workflows have been removed as they are no longer needed. A new linters workflow has been added to run all linters and improve code quality. This new workflow runs on every push and pull request to the master branch. It sets the TERM environment variable, checks out the code, sets up Python 3.10, installs dependencies, and runs all linters using the `make lint` command.

* ✨ chore(github): add Ansible-lint, Flake8, and Yamllint workflows to improve code quality
The Ansible-lint, Flake8, and Yamllint workflows have been added to the project to improve code quality. These workflows will run on every push and pull request to the master branch. The Ansible-lint workflow checks for syntax errors and best practices in Ansible code, the Flake8 workflow checks for PEP8 compliance and other code quality issues in Python code, and the Yamllint workflow checks for syntax errors and best practices in YAML files. The README.md file has also been updated to reflect the changes in the project.

* 🔀 chore(README.md): reformat index and cluster types sections for better readability
The index and cluster types sections were reformatted to improve readability and make it easier to navigate the document. The changes include removing unnecessary characters, reordering the sections, and changing the formatting of the sub-sections.

🎨 style(README.md): reformat table for better readability
The table was reformatted to improve readability and make it easier to compare the test results across different distributions.

* 📝 chore(README.md): reformat table of test results for better readability
The table of test results has been reformatted to improve readability. The table now has a cleaner look and is easier to read.