Add a GitHub Action for Docker (make docker-build) (#365)
* 🆕 chore(gitpod.yml): add Gitpod workflow to run tests on push and pull requests
This commit adds a new Gitpod workflow to run tests on push and pull requests on the master branch. The workflow sets up Python 3.10, installs dependencies, and runs Docker tests using the `make docker-tests` command. The `TERM` environment variable is also set to `xterm` to avoid issues with the terminal.

* 🆕 chore(docker.yml): add Docker workflow to build and test Docker images
This commit adds a new file, `.github/workflows/docker.yml`, which defines a workflow to build and test Docker images. The workflow is triggered on push and pull requests to the master branch. The workflow runs on Ubuntu latest and sets up Python 3.10. It then installs dependencies and runs Docker tests using the `make docker-tests` command.