feat: add tests for Ansible Verifier (#367)
* 🚀 feat(molecule): add Postgres and Patroni tests
✨ feat(molecule): add etcd health check test
🐛 fix(molecule): enable verifier in molecule config
The molecule configuration file was updated to enable the verifier. A new group_vars file was added to set the default Postgres database name, user, and password. Tests were added to check if PostgreSQL is running, listening on the default port, and if the database exists. A test was added to check the replication status of PostgreSQL. Tests were added to check the status of Patroni and etcd. These tests will help ensure that the application is running correctly and that the infrastructure is properly configured.

* 🐛 fix(verify.yml): reorder fileglob

Co-authored-by: Vitaliy Kukharik <37010174+vitabaks@users.noreply.github.com>

* 🔥 chore(all.yml): remove all.yml file
✨ feat(verify.yml): include all tests and remove replica tests
🐛 fix(postgres.yml): use patroni_superuser_username and patroni_superuser_password instead of postgres_user and postgres_password
🐛 fix(postgres.yml): fix typo in task name
The all.yml file was removed as it was no longer needed. The verify.yml file was updated to include all tests and remove replica tests. The postgres.yml file was updated to use patroni_superuser_username and patroni_superuser_password instead of postgres_user and postgres_password. A typo in the task name was also fixed.

* 🔥 chore(postgres.yml): remove redundant test to check if databases exist
The test to check if the databases exist is redundant as it is already being checked in the `molecule/default/converge.yml` playbook. Removing this test will reduce the time taken to run the tests and improve the overall performance of the test suite.

* 🐛 fix(verify.yml): fix indentation of file path in include_vars task
The indentation of the file path in the include_vars task was fixed to be consistent with the rest of the file.

* 🐛 fix(postgres.yml): change database name from postgres to template1
The database name was changed from postgres to template1 to avoid issues with the postgres database being used as a template database.

---------

Co-authored-by: Vitaliy Kukharik <37010174+vitabaks@users.noreply.github.com>