Revise ansible-lint to incorporate incremental changes for future adaptability (#368)
* 🔧 chore(ansible-lint.yml): update skip_list to ignore additional yaml rules
The skip_list in ansible-lint.yml has been updated to ignore additional yaml rules. The rules that have been added are yaml[line-length], yaml[octal-values], yaml[truthy], and yaml[trailing-spaces]. These rules have been added to improve the quality of the code by ensuring that it adheres to best practices.

* 🔍 chore(ansible-lint.yml): add role-name[path] to skip_list
The role-name[path] rule is added to the skip_list in order to ignore the warning generated by ansible-lint for the role-name[path] rule. This rule is skipped because it is not applicable to the project's directory structure.

* 🔧 chore(ansible-lint.yml): update skip_list to include fqcn[action-core] instead of fqcn-builtins
The fqcn-builtins rule has been deprecated and replaced with fqcn[action-core]. This commit updates the skip_list in ansible-lint.yml to reflect this change.

* 🔥 chore(ansible-lint.yml): remove template-instead-of-copy from skip_list
The template-instead-of-copy rule was removed from the skip_list in the ansible-lint.yml configuration file. This rule was previously skipped, but it is now enabled to ensure that templated files use the template keyword instead of copy.

* 🔥 chore(.gitignore): add __pycache__/ to ignore Python cache files
The __pycache__/ directory is added to the .gitignore file to prevent Python cache files from being committed to the repository. This improves the cleanliness of the repository and reduces the size of the repository.

* 🔧 chore(ansible-lint.yml): update skip_list to include fqcn[action] rule
The fqcn[action] rule is now included in the skip_list to avoid it being flagged as an error. This rule is skipped because it is not applicable to the project's current configuration.

* 🔥 chore(ansible-lint.yml): remove var-naming[no-role-prefix] rule from skip_list
The var-naming[no-role-prefix] rule was removed from the skip_list as it is no longer needed.

* 🔧 chore(ansible-lint.yml): update skip_list to reflect new rules and TODOs
The ansible-lint.yml file has been updated to reflect new rules and TODOs. The skip_list has been updated to include new rules such as no-free-form, no-relative-paths, and key-order[task]. Additionally, some rules have been marked as TODOs, such as risky-file-permissions and jinja[invalid]. Finally, the order of the skip_list has been updated to group similar rules together.

* 🔍 chore(ansible-lint.yml): add run-once[task] to skip_list
The run-once[task] rule is added to the skip_list in the ansible-lint configuration file. This rule is skipped during linting to avoid warnings about tasks that are intended to run only once.